Create cell_type_counting_and_area.m
Generally, this code loads a transverse optical section .svg, the user made using Abode Illustrator, then it meshes the cell walls and applies a 2x radially thicker outer cell wall and plots this visually. This  visualization is useful as if the .svg file needs editing it will show internal cells which need to be edited. Next, the code allows the user to select cell type by clicking each cell. After cell type selection, the number of each cell type and area for each cell is calculate.